gnome extension: Give priority to local environment
When creating the environment for gnome shell, the order in
which the variables from the SDK should be added was reversed:
the SDK ones should be at the end, and the new elements should
be at the beginning. This allows to priorityze the local
changes over the generic elements in the SDK.

This was commented in Mattermost.